Create what's new and add printer dialogs dynamically
CURA-12671
